backend/pyproject.toml
[project]
name = "hotpot-backend"
version = "0.1.0"
description = "Hotpot Tech Feed — backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "feedparser>=6.0",
  "openai>=1.30",
  "qdrant-client>=1.9",
  "rapidfuzz>=3.6",
  "pyyaml>=6.0",
  "python-dateutil>=2.9",
  "tldextract>=5.1",
  "celery[redis]>=5.4",
  "redis>=5.0",
  "click>=8.1",
  "structlog>=24.1",
]

[project.optional-dependencies]
embeddings = [
  "sentence-transformers>=3.0",
  "torch>=2.1",
]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
]

[project.scripts]
hotpot = "app.cli:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
